# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Coffee Log API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hoppers:
    get:
      operationId: listHoppers
      tags: [hoppers]
      responses:
        "200":
          description: Hopper states
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Hopper" }
    post:
      operationId: createHopper
      tags: [hoppers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HopperInput" }
      responses:
        "201":
          description: Created hopper state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hopper" }
  /hoppers/{id}:
    patch:
      operationId: updateHopper
      tags: [hoppers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HopperUpdate" }
      responses:
        "200":
          description: Updated hopper state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hopper" }
  /hoppers/import-csv:
    post:
      operationId: importHoppersCsv
      tags: [hoppers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CsvImportInput" }
      responses:
        "200":
          description: Hopper import result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CsvImportResult" }
  /hopper-range-baselines:
    get:
      operationId: listHopperRangeBaselines
      tags: [hoppers]
      responses:
        "200":
          description: Hopper range baselines
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HopperRangeBaseline" }
    post:
      operationId: createHopperRangeBaseline
      tags: [hoppers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HopperRangeBaselineInput" }
      responses:
        "201":
          description: Created baseline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HopperRangeBaseline" }
  /hopper-range-baselines/import-csv:
    post:
      operationId: importHopperRangeBaselinesCsv
      tags: [hoppers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CsvImportInput" }
      responses:
        "200":
          description: Hopper range baseline import result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CsvImportResult" }
components:
  schemas:
...
    CsvImportInput:
      type: object
      required: [csvText]
      properties:
        csvText:
          type: string
    Hopper:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        bagId: { type: ["integer", "null"] }
        startingBeans: { type: ["number", "null"] }
        isActive: { type: boolean }
        hopperMass: { type: ["number", "null"], readOnly: true }
        hopperPercent: { type: ["number", "null"], readOnly: true }
        shotsLeftEstimate: { type: ["number", "null"], readOnly: true }
        phase: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    HopperInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        bagId: { type: ["integer", "null"] }
        startingBeans: { type: ["number", "null"] }
        isActive: { type: boolean }
        phase: { type: string }
        notes: { type: string }
    HopperUpdate:
      type: object
      properties:
        name: { type: string }
        bagId: { type: ["integer", "null"] }
        startingBeans: { type: ["number", "null"] }
        isActive: { type: boolean }
        phase: { type: string }
        notes: { type: string }
    HopperRangeBaseline:
      type: object
      required: [id, hopperRange, createdAt]
      properties:
        id: { type: integer }
        hopperRange: { type: string }
        baselineOutputAdjustedDate: { type: ["string", "null"] }
        baselineOutputStatus: { type: ["string", "null"] }
        baselineOutput: { type: ["number", "null"] }
        avgInitialOutput: { type: ["number", "null"], readOnly: true }
        observationCount: { type: ["integer", "null"], readOnly: true }
        createdAt: { type: string }
    HopperRangeBaselineInput:
      type: object
      required: [hopperRange]
      properties:
        hopperRange: { type: string }
        baselineOutputAdjustedDate: { type: ["string", "null"] }
        baselineOutputStatus: { type: string }
        baselineOutput: { type: ["number", "null"] }
    CsvImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
...