# lib/api-zod/src/generated/types/hopper.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Coffee Log API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Hopper {
  id: number;
  name: string;
  /** @nullable */
  bagId?: number | null;
  /** @nullable */
  startingBeans?: number | null;
  isActive: boolean;
  /** @nullable */
  readonly hopperMass?: number | null;
  /** @nullable */
  readonly hopperPercent?: number | null;
  /** @nullable */
  readonly shotsLeftEstimate?: number | null;
  /** @nullable */
  phase?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}