# lib/api-zod/src/generated/types/hopperInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Coffee Log API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HopperInput {
  name: string;
  /** @nullable */
  bagId?: number | null;
  /** @nullable */
  startingBeans?: number | null;
  isActive?: boolean;
  phase?: string;
  notes?: string;
}