# lib/api-zod/src/generated/types/hopperUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Coffee Log API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HopperUpdate {
  name?: string;
  /** @nullable */
  bagId?: number | null;
  /** @nullable */
  startingBeans?: number | null;
  isActive?: boolean;
  phase?: string;
  notes?: string;
}